# specs/001-mini-twitter-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SocialMedia HTTP Contracts
  version: 1.0.0
  description: |
    Versioned public and private-network REST contracts. Public `/api/v1` routes are exposed
    by Gateway; `/internal/v1` routes are never routed by Gateway. Every response returns
    `X-Correlation-Id`. Protected operations derive the actor from JWT `sub`.
servers:
  - url: http://localhost:8080
    description: Local API Gateway
...
paths:
  /api/v1/auth/register:
    post:
      tags: [User]
      operationId: register
      summary: Register one account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [User]
      operationId: login
      summary: Authenticate without revealing which credential was wrong
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Signed credential issued
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    CorrelationId:
      description: Request correlation ID accepted or generated by Gateway
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    BadRequest:
      description: Validation or malformed cursor failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, malformed, or invalid credential
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: Case-insensitive username or email conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    UuidV7:
      type: string
      format: uuid
      description: RFC 9562 UUID version 7 generated by the owning service
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, username, displayName, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[A-Za-z0-9_]+$'
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        bio:
          type: string
          maxLength: 160
        password:
          type: string
          format: password
          minLength: 12
          description: Maximum 72 UTF-8 bytes; never logged or returned
    LoginRequest:
      type: object
      additionalProperties: false
      required: [login, password]
      properties:
        login:
          type: string
          description: Case-insensitive username or email
        password:
          type: string
          format: password
    TokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresAt]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresAt:
          type: string
          format: date-time
    AccountView:
      type: object
      additionalProperties: false
      required: [id, email, username, displayName, bio, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/UuidV7'
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        correlationId:
          type: string
security:
  - bearerAuth: []